refactor: behaviours to templates
